# backend/ent/admin/types.go
// Code generated by ent, DO NOT EDIT.
package admin

import "time"

type AnswerChoice struct {
	ChoiceText string `form:"choice_text"`
	IsCorrect  bool   `form:"is_correct"`
}

func (e *AnswerChoice) GetName() string {
	return NameAnswerChoice
}

func (e *AnswerChoice) GetSchema() []*FieldSchema {
	return fieldsAnswerChoice
}

type Category struct {
	Name        string  `form:"name"`
	Slug        string  `form:"slug"`
	Description *string `form:"description"`
	Image       *string `form:"image"`
}

func (e *Category) GetName() string {
	return NameCategory
}

func (e *Category) GetSchema() []*FieldSchema {
	return fieldsCategory
}

type Certificate struct {
	UniqueID  string     `form:"unique_id"`
	IssueDate *time.Time `form:"issue_date"`
}

func (e *Certificate) GetName() string {
	return NameCertificate
}

func (e *Certificate) GetSchema() []*FieldSchema {
	return fieldsCertificate
}

type Course struct {
	Title            string     `form:"title"`
	Slug             string     `form:"slug"`
	ShortDescription string     `form:"short_description"`
	Description      string     `form:"description"`
	WhatYouWillLearn *string    `form:"what_you_will_learn"`
	Requirements     *string    `form:"requirements"`
	TargetAudience   *string    `form:"target_audience"`
	Thumbnail        *string    `form:"thumbnail"`
	PromoVideoLink   *string    `form:"promo_video_link"`
	Price            *float64   `form:"price"`
	IsFree           bool       `form:"is_free"`
	IsPublished      bool       `form:"is_published"`
	Level            *string    `form:"level"`
	Duration         *string    `form:"duration"`
	TotalLectures    *int       `form:"total_lectures"`
	AverageRating    *float64   `form:"average_rating"`
	NumberOfReviews  *int       `form:"number_of_reviews"`
	CreatedAt        *time.Time `form:"created_at"`
	UpdatedAt        *time.Time `form:"updated_at"`
}

func (e *Course) GetName() string {
	return NameCourse
}

func (e *Course) GetSchema() []*FieldSchema {
	return fieldsCourse
}

type Enrollment struct {
	EnrolledAt  *time.Time `form:"enrolled_at"`
	CompletedAt *time.Time `form:"completed_at"`
	Progress    *float64   `form:"progress"`
	IsCompleted bool       `form:"is_completed"`
}

func (e *Enrollment) GetName() string {
	return NameEnrollment
}

func (e *Enrollment) GetSchema() []*FieldSchema {
	return fieldsEnrollment
}

type FAQ struct {
	Question    string     `form:"question"`
	Answer      string     `form:"answer"`
	IsPublished bool       `form:"is_published"`
	Order       *int       `form:"order"`
	CreatedAt   *time.Time `form:"created_at"`
	UpdatedAt   *time.Time `form:"updated_at"`
}

func (e *FAQ) GetName() string {
	return NameFAQ
}

func (e *FAQ) GetSchema() []*FieldSchema {
	return fieldsFAQ
}

type Lesson struct {
	Title     string     `form:"title"`
	Slug      string     `form:"slug"`
	Content   *string    `form:"content"`
	VideoLink *string    `form:"video_link"`
	Order     int        `form:"order"`
	IsPreview bool       `form:"is_preview"`
	CreatedAt *time.Time `form:"created_at"`
	UpdatedAt *time.Time `form:"updated_at"`
}

func (e *Lesson) GetName() string {
	return NameLesson
}

func (e *Lesson) GetSchema() []*FieldSchema {
	return fieldsLesson
}

type PasswordToken struct {
	Token     *string    `form:"token"`
	UserID    int        `form:"user_id"`
	CreatedAt *time.Time `form:"created_at"`
}

func (e *PasswordToken) GetName() string {
	return NamePasswordToken
}

func (e *PasswordToken) GetSchema() []*FieldSchema {
	return fieldsPasswordToken
}

type Question struct {
	QuestionText string  `form:"question_text"`
	QuestionType *string `form:"question_type"`
	Order        *int    `form:"order"`
}

func (e *Question) GetName() string {
	return NameQuestion
}

func (e *Question) GetSchema() []*FieldSchema {
	return fieldsQuestion
}

type Quiz struct {
	Title        string     `form:"title"`
	Description  *string    `form:"description"`
	PassingScore *int       `form:"passing_score"`
	CreatedAt    *time.Time `form:"created_at"`
	UpdatedAt    *time.Time `form:"updated_at"`
}

func (e *Quiz) GetName() string {
	return NameQuiz
}

func (e *Quiz) GetSchema() []*FieldSchema {
	return fieldsQuiz
}

type Review struct {
	Rating     int        `form:"rating"`
	Comment    *string    `form:"comment"`
	CreatedAt  *time.Time `form:"created_at"`
	IsApproved bool       `form:"is_approved"`
}

func (e *Review) GetName() string {
	return NameReview
}

func (e *Review) GetSchema() []*FieldSchema {
	return fieldsReview
}

type User struct {
	Username             string     `form:"username"`
	FirstName            *string    `form:"first_name"`
	LastName             *string    `form:"last_name"`
	Email                string     `form:"email"`
	Password             *string    `form:"password"`
	Role                 *string    `form:"role"`
	Bio                  *string    `form:"bio"`
	ProfilePicture       *string    `form:"profile_picture"`
	DateOfBirth          *time.Time `form:"date_of_birth"`
	Gender               *string    `form:"gender"`
	ContactNumber        *string    `form:"contact_number"`
	Address              *string    `form:"address"`
	Country              *string    `form:"country"`
	IsEmailVerified      bool       `form:"is_email_verified"`
	HighestQualification *string    `form:"highest_qualification"`
	Institution          *string    `form:"institution"`
	Skills               *string    `form:"skills"`
	LinkedinProfile      *string    `form:"linkedin_profile"`
	GithubProfile        *string    `form:"github_profile"`
	InstructorRating     *float64   `form:"instructor_rating"`
	TotalReviews         *int       `form:"total_reviews"`
	IsActiveUser         bool       `form:"is_active_user"`
	LastActivity         *time.Time `form:"last_activity"`
	LoginAddress         *string    `form:"login_address"`
	TwoFactorEnabled     bool       `form:"two_factor_enabled"`
	IsStaff              bool       `form:"is_staff"`
	IsActive             bool       `form:"is_active"`
	IsSuperuser          bool       `form:"is_superuser"`
	LastLogin            *time.Time `form:"last_login"`
	DateJoined           *time.Time `form:"date_joined"`
	Verified             bool       `form:"verified"`
	Admin                bool       `form:"admin"`
	CreatedAt            *time.Time `form:"created_at"`
	OtpCode              *string    `form:"otp_code"`
	OtpExpiresAt         *time.Time `form:"otp_expires_at"`
}

func (e *User) GetName() string {
	return NameUser
}

func (e *User) GetSchema() []*FieldSchema {
	return fieldsUser
}

type UserQuizAttempt struct {
	Score         *float64   `form:"score"`
	Passed        bool       `form:"passed"`
	AttemptNumber *int       `form:"attempt_number"`
	SubmittedAt   *time.Time `form:"submitted_at"`
}

func (e *UserQuizAttempt) GetName() string {
	return NameUserQuizAttempt
}

func (e *UserQuizAttempt) GetSchema() []*FieldSchema {
	return fieldsUserQuizAttempt
}

type EntityType interface {
	GetName() string
	GetSchema() []*FieldSchema
}

var entityTypes = []EntityType{
	&AnswerChoice{},
	&Category{},
	&Certificate{},
	&Course{},
	&Enrollment{},
	&FAQ{},
	&Lesson{},
	&PasswordToken{},
	&Question{},
	&Quiz{},
	&Review{},
	&User{},
	&UserQuizAttempt{},
}

type EntityList struct {
	Columns     []string
	Entities    []EntityValues
	Page        int
	HasNextPage bool
}

type EntityValues struct {
	ID     int
	Values []string
}

type HandlerConfig struct {
	ItemsPerPage int
	PageQueryKey string
	TimeFormat   string
}

func GetEntityTypes() []EntityType {
	return entityTypes
}